@@ -10,21 +10,21 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "build-schematics": "tsc -p projects/ngx-metro/tsconfig.schematics.json --outDir dist/ngx-metro/schematics && cpx dist/ngx-metro/schematics/src/schematics/** dist/ngx-metro/schematics && cpx projects/ngx-metro/src/schematics/**/{collection.json,schema.json,files/**} dist/ngx-metro/schematics",
-    "build-lib": "npm run build -- ngx-metro --prod && npm run build-schematics && cpx ./README.md ./dist/ngx-metro",
+    "build-lib": "npm run build -- ngx-metro --configuration production && npm run build-schematics && cpx ./README.md ./dist/ngx-metro",
     "publish-lib": "npm run build-lib && cd ./dist/ngx-metro && npm publish",
     "publish-doc": "ng run ngx-metro-demo:deploy"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.4.0",
-    "@angular/common": "~13.4.0",
-    "@angular/compiler": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "~13.4.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
-    "@angular/service-worker": "~13.4.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "core-js": "^2.5.4",
     "highlight.js": "^9.15.10",
     "linq4js": "^2.2.2",
@@ -36,11 +36,11 @@
     "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.11",
-    "@angular-devkit/build-ng-packagr": "~0.1002.4",
-    "@angular/cli": "^13.3.11",
-    "@angular/compiler-cli": "13.4.0",
-    "@angular/language-service": "~13.4.0",
+    "@angular-devkit/build-angular": "^14.2.12",
+    "@angular-devkit/build-ng-packagr": "^0.1002.4",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@angularclass/hmr": "^2.1.3",
     "@babel/core": "^7.6.0",
     "@types/jasmine": "~2.8.8",
@@ -58,7 +58,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~6.1.0",
