@@ -1,4 +1,4 @@
-import {AbstractControl, AbstractControlOptions, AsyncValidatorFn, FormGroup, ValidatorFn} from '@angular/forms';
+import {AbstractControl, AbstractControlOptions, AsyncValidatorFn, UntypedFormGroup, ValidatorFn} from '@angular/forms';
 import {Observable} from 'rxjs';
 
 export interface FormStyle {
@@ -8,7 +8,7 @@ export interface FormStyle {
   createError?: (error: string) => string|Observable<string>;
 }
 
-export class MetroFormGroup extends FormGroup {
+export class MetroFormGroup extends UntypedFormGroup {
   constructor(public name: string,
               controls: { [key: string]: AbstractControl; },
               validatorOrOpts?: ValidatorFn | ValidatorFn[] | AbstractControlOptions | null,
