@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {FormControl, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormControl, UntypedFormGroup, Validators} from '@angular/forms';
 import { MetroFormGroup } from 'ngx-metro';
 import {CustomGroupValidator} from './custom-group-validator';
 
@@ -10,16 +10,16 @@ import {CustomGroupValidator} from './custom-group-validator';
 })
 export class FormComponent implements OnInit {
 
-  public formGroup: FormGroup;
+  public formGroup: UntypedFormGroup;
 
   constructor() {
 
     this.formGroup = new MetroFormGroup('reactive_demo', {
-      email: new FormControl('', [Validators.required, Validators.email, Validators.minLength(3)]),
-      password: new FormControl('', [Validators.required]),
-      number: new FormControl(10, [Validators.min(5), Validators.max(30)]),
-      password_confirm: new FormControl('', [Validators.required]),
-      remember: new FormControl(false, [Validators.requiredTrue])
+      email: new UntypedFormControl('', [Validators.required, Validators.email, Validators.minLength(3)]),
+      password: new UntypedFormControl('', [Validators.required]),
+      number: new UntypedFormControl(10, [Validators.min(5), Validators.max(30)]),
+      password_confirm: new UntypedFormControl('', [Validators.required]),
+      remember: new UntypedFormControl(false, [Validators.requiredTrue])
     }, [CustomGroupValidator.sameValues('password', 'password_confirm')]);
   }
 
